requirements.txt
fastapi
uvicorn
requests
python-dotenv
jinja2
python-multipart
pydantic